+++ config/config.go
@@ -49,10 +49,10 @@ type SectionAPI struct {
 
 // SectionGitHub is a sub section of config.
 type SectionGitHub struct {
-	AppID         int            `yaml:"app_id"`
-	Secret        string         `yaml:"secret"`
-	PrivateKey    string         `yaml:"private_key"`
-	Installations map[string]int `yaml:"installations"`
+	AppID         int64            `yaml:"app_id"`
+	Secret        string           `yaml:"secret"`
+	PrivateKey    string           `yaml:"private_key"`
+	Installations map[string]int64 `yaml:"installations"`
 }
 
 // SectionLog is a sub section of config.
@@ -106,7 +106,7 @@ func BuildDefaultConf() Config {
 	conf.GitHub.AppID = 0
 	conf.GitHub.Secret = ""
 	conf.GitHub.PrivateKey = ""
-	conf.GitHub.Installations = make(map[string]int)
+	conf.GitHub.Installations = make(map[string]int64)
 
 	// Log
 	conf.Log.Format = "string"
